add FilterBaseNoop so that subclasses are not forced to use setters
