+++ java/org/apache/catalina/filters/CorsFilter.java
@@ -28,7 +28,6 @@ import java.util.Locale;
 import java.util.Set;
 
 import javax.servlet.FilterChain;
-import javax.servlet.GenericFilter;
 import javax.servlet.ServletException;
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
@@ -75,7 +74,7 @@ import org.apache.tomcat.util.res.StringManager;
  * @see <a href="http://www.w3.org/TR/cors/">CORS specification</a>
  *
  */
-public class CorsFilter extends FilterBase {
+public class CorsFilter extends FilterBaseNoop {
 
     private static final long serialVersionUID = 1L;
     private static final Log log = LogFactory.getLog(CorsFilter.class);
