+++ java/org/apache/catalina/filters/FilterBase.java
@@ -19,24 +19,21 @@ package org.apache.catalina.filters;
 import java.util.Enumeration;
 
 import javax.servlet.FilterConfig;
-import javax.servlet.GenericFilter;
 import javax.servlet.ServletException;
 
-import org.apache.juli.logging.Log;
 import org.apache.tomcat.util.IntrospectionUtils;
 import org.apache.tomcat.util.res.StringManager;
 
 /**
- * Base class for filters that provides generic initialisation and a simple
- * no-op destruction.
+ * Base class for filters that provides generic initialisation via setter methods that
+ * match config init parameters, and a simple no-op destruction.
  *
  * @author xxd
  *
  */
-public abstract class FilterBase extends GenericFilter {
+public abstract class FilterBase extends FilterBaseNoop {
 
     protected static final StringManager sm = StringManager.getManager(FilterBase.class);
-
 
     /**
      * Iterates over the configuration parameters and either logs a warning,
@@ -81,38 +78,4 @@ public abstract class FilterBase extends GenericFilter {
         return false;
     }
 
-    /**
-     * This method returns the parameter's value if it exists, or defaultValue if not.
-     *
-     * @param name - The parameter's name
-     * @param defaultValue - The default value to return if the parameter does not exist
-     * @return The parameter's value or the default value if the parameter does not exist
-     */
-    public String getInitParameter(String name, String defaultValue){
-
-        String value = getInitParameter(name);
-
-        if (value == null)
-            return defaultValue;
-
-        return value;
-    }
-
-    /**
-     * Sub-classes can return a Log that will be used to write log entries
-     * @return The default implementation returns null which means that no log entries will be written
-     */
-    protected Log getLogger(){
-        return null;
-    }
-
-    /**
-     * Logs the message at warn level if a logger exists
-     * @param message The warning message to be logged
-     */
-    protected void warn(Object message){
-        Log log = getLogger();
-        if (log != null)
-            log.warn(message);
-    }
 }
